rename parking spaces dir and file
