+++ packages/frontend/src/App.tsx
@@ -19,7 +19,7 @@ import GraphikRegular from '../assets/Graphik-Regular.woff2'
 import GraphikBold from '../assets/Graphik-Bold.woff2'
 import Login from './pages/Login/Login'
 import MaterialChoiceDetails from './pages/MaterialChoiceDetails/MaterialChoiceDetails'
-import ParkingSpaces from './pages/Parkingspaces'
+import ParkingSpaces from './pages/ParkingSpaces'
 import ParkingSpace from './pages/ParkingSpace'
 import { ParkingSpaceLoading } from './pages/ParkingSpace/components'
